Extract shared GP test fixtures into conftest
Move mock_app and manager fixtures from test_manager.py into a shared
conftest.py so new test modules (real-frame tests, future E2E cases)
can reuse them without duplication.
