+++ tests/zgp/test_manager.py
@@ -2,9 +2,7 @@
 
 from __future__ import annotations
 
-import asyncio
-
-from tests.async_mock import AsyncMock, MagicMock, Mock
+from tests.async_mock import AsyncMock, MagicMock
 
 import pytest
 
@@ -19,27 +17,6 @@ from zigpy.zgp.types import (
     SecurityKeyType,
     SecurityLevel,
 )
-
-
-@pytest.fixture
-def mock_app() -> MagicMock:
-    """Create a mock ControllerApplication."""
-    app = MagicMock()
-    app.state.node_info.ieee = t.EUI64.convert("00:11:22:33:44:55:66:77")
-    app.state.node_info.nwk = t.NWK(0x0000)
-    app.state.network_info.channel = 15
-    app.send_packet = AsyncMock()
-    app.listener_event = Mock()
-    app.create_task = Mock(
-        side_effect=lambda coro, name=None: asyncio.ensure_future(coro)
-    )
-    return app
-
-
-@pytest.fixture
-def manager(mock_app: MagicMock) -> GreenPowerManager:
-    """Create a GreenPowerManager with mock app."""
-    return GreenPowerManager(mock_app)
 
 
 def _make_gp_notification_packet(
